+++ R/grec-internal.R
@@ -1,48 +1,62 @@
 checkArgs_df_matrix <- function(x){
-  # Error messages
-  msg1 <- "'x' must be a numeric matrix with environmental data. See help(getGradients)."
-  msg2 <- "There is not any valid values for 'x'"
-
   # Check if x is a valid numerical matrix
   if(!is.numeric(x)){
-    stop(msg1)
+    cli_abort(
+      message = c(
+        "{.code x} must be a numeric matrix with environmental data.",
+        "i" = "See {.run help(getGradients)}."
+      )
+    )
   }
 
   # Check if x is a valid numerical matrix
   if(sum(is.na(as.numeric(x))) == prod(dim(x))){
-    stop(msg2)
+    cli_abort(
+      message = c(
+        "There is not any valid values for {.code x}."
+      )
+    )
   }
 
   invisible()
 }
 
 checkArgs_df_array <- function(x){
-  # Error messages
-  msg1 <- "'x' must be a numeric array with environmental data. See help(getGradients)."
-  msg2 <- "There is not any valid values for 'x'"
-
   # Check if x is a valid numerical matrix
   if(!is.numeric(x)){
-    stop(msg1)
+    if(!is.numeric(x)){
+      cli_abort(
+        message = c(
+          "{.code x} must be a numeric array with environmental data.",
+          "i" = "See {.run help(getGradients)}."
+        )
+      )
+    }
   }
 
   # Check if x is a valid numerical matrix
   if(sum(is.na(as.numeric(x))) == prod(dim(x))){
-    stop(msg2)
+    cli_abort(
+      message = c(
+        "There is not any valid values for {.code x}."
+      )
+    )
   }
 
   invisible()
 }
 
 checkArgs_df_list <- function(x){
-  # Error messages
-  msg1 <- "'x' must be a XYZ list containing environmental map info (whether a matrix or an array). See help(getGradients)."
-
   # Check if x is a list with 'x', 'y', 'z' dimensions, where z is a numeric matrix/array
   index <- (length(x) == 3 && all(is.element(c("x", "y", "z"), names(x))) && is.numeric(x$x) && is.numeric(x$y) &&
               (is.matrix(x$z) || is.array(x$z)) && is.numeric(x$z))
   if(!index){
-    stop(msg1)
+    cli_abort(
+      message = c(
+        "{.code x} must be a XYZ list containing environmental map info (whether a matrix or an array).",
+        "i" = "See {.run help(getGradients)}."
+      )
+    )
   }
 
   invisible()
@@ -51,7 +65,11 @@ checkArgs_df_list <- function(x){
 checkArgs_df_RasterLayer <- function(x){
 
   if(nlayers(x) < 1){
-    stop("RasterLayer object must have at least 1 layer.")
+    cli_abort(
+      message = c(
+        "RasterLayer object must have at least 1 layer."
+      )
+    )
   }
   invisible()
 }
@@ -59,7 +77,11 @@ checkArgs_df_RasterLayer <- function(x){
 checkArgs_df_SpatRaster <- function(x){
 
   if(nlyr(x) < 1){
-    stop("SpatRaster object must have at least 1 layer.")
+    cli_abort(
+      message = c(
+        "SpatRaster object must have at least 1 layer."
+      )
+    )
   }
 
   invisible()
@@ -77,17 +99,29 @@ checkArgs_prevs <- function(allArgs, type){
   # Check name of method
   methodList <- c("belkinoreilly2009", "medianfilter", "agenbag20031", "agenbag20032")
   if(!is.element(method, methodList)){
-    stop("No valid method for gradient calculation.")
+    cli_abort(
+      message = c(
+        "No valid method for gradient calculation."
+      )
+    )
   }
 
   # Check 'intermediate'
   if(length(intermediate) != 1 || !is.logical(intermediate)){
-    stop("'intermediate' must be a single logical value.")
+    cli_abort(
+      message = c(
+        "{.code intermediate} must be a single logical value."
+      )
+    )
   }
 
   # Check 'ConvolNormalization'
   if(length(ConvolNormalization) != 1 || !is.logical(ConvolNormalization)){
-    stop("'ConvolNormalization' must be a single logical value.")
+    cli_abort(
+      message = c(
+        "{.code ConvolNormalization} must be a single logical value."
+      )
+    )
   }
 
   invisible()