+++ R/grec-package.R
@@ -7,13 +7,32 @@
 
 ## usethis namespace: start
 #' @importFrom abind abind
-#' @importFrom imagine contextualMF convolution2D medianFilter agenbagFilters
-#' @importFrom lifecycle deprecate_soft deprecate_warn badge
+#' @importFrom cli cli_abort
+#' @importFrom imagine agenbagFilters
+#' @importFrom imagine contextualMF
+#' @importFrom imagine convolution2D
+#' @importFrom imagine medianFilter
+#' @importFrom lifecycle badge
+#' @importFrom lifecycle deprecate_soft
+#' @importFrom lifecycle deprecate_warn
 #' @importFrom lifecycle deprecated
-#' @importFrom raster raster values nlayers
-#' @importFrom terra rast values nlyr crs ext varnames origin 'values<-'
+#' @importFrom raster nlayers
+#' @importFrom raster raster
+#' @importFrom raster values
+#' @importFrom terra 'values<-'
+#' @importFrom terra as.array
+#' @importFrom terra crs
+#' @importFrom terra ext
+#' @importFrom terra nlyr
+#' @importFrom terra origin
+#' @importFrom terra rast
+#' @importFrom terra values
+#' @importFrom terra varnames
 #' @importFrom utils modifyList
 #' @importMethodsFrom raster "["
-#' @importMethodsFrom terra as.matrix varnames<- origin<- ext<-
+#' @importMethodsFrom terra as.matrix
+#' @importMethodsFrom terra ext<-
+#' @importMethodsFrom terra origin<-
+#' @importMethodsFrom terra varnames<-
 ## usethis namespace: end
 NULL