+++ R/list-class.R
@@ -8,8 +8,13 @@ getGradients.list <- function(x, method = "BelkinOReilly2009",
   checkArgs_df_list(x = x)
 
   # Apply method
-  x$z <- getGradients(x = x$z, method = method, intermediate = intermediate,
-                      checkPrevs = FALSE, ...)
+  x$z <- getGradients(
+    x = x$z,
+    method = method,
+    intermediate = intermediate,
+    checkPrevs = FALSE,
+    ...
+  )
 
   x
 }